+++ Source/ThirdParty/tomcat/bin/tool-wrapper.sh
@@ -18,7 +18,7 @@
 # -----------------------------------------------------------------------------
 # Wrapper script for command line tools
 #
-# Environment Variable Prequisites
+# Environment Variable Prerequisites
 #
 #   CATALINA_HOME May point at your Catalina "build" directory.
 #
@@ -30,7 +30,7 @@
 #   JAVA_OPTS     (Optional) Java runtime options used when the "start",
 #                 "stop", or "run" command is executed.
 #
-# $Id: tool-wrapper.sh 947596 2010-05-24 10:50:27Z kkolinko $
+# $Id: tool-wrapper.sh 1040555 2010-11-30 15:00:25Z rjung $
 # -----------------------------------------------------------------------------
 
 # OS specific support.  $var _must_ be set to either true or false.